rename ZSchedule to Schedule
